acpapp/fastapi/requirements.txt
fastapi[standard]
uvicorn
databases[asyncpg]
pydantic
bcrypt
python-jose